Split fulfillment tests
