+++ saleor/graphql/order/tests/mutations/test_fulfillment_update_tracking.py
@@ -1,0 +1,88 @@
+from unittest.mock import ANY, patch
+
+import graphene
+
+from ....tests.utils import get_graphql_content
+
+FULFILLMENT_UPDATE_TRACKING_QUERY = """
+    mutation updateFulfillment(
+        $id: ID!
+        $tracking: String
+        $notifyCustomer: Boolean
+    ) {
+        orderFulfillmentUpdateTracking(
+            id: $id
+            input: { trackingNumber: $tracking, notifyCustomer: $notifyCustomer }
+        ) {
+            fulfillment {
+                trackingNumber
+            }
+        }
+    }
+"""
+
+
+@patch("saleor.plugins.manager.PluginsManager.notify")
+def test_fulfillment_update_tracking(
+    send_fulfillment_update_mock,
+    staff_api_client,
+    fulfillment,
+    permission_manage_orders,
+):
+    query = FULFILLMENT_UPDATE_TRACKING_QUERY
+    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
+    tracking = "stationary tracking"
+    variables = {"id": fulfillment_id, "tracking": tracking}
+    response = staff_api_client.post_graphql(
+        query, variables, permissions=[permission_manage_orders]
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["orderFulfillmentUpdateTracking"]["fulfillment"]
+    assert data["trackingNumber"] == tracking
+    send_fulfillment_update_mock.assert_not_called()
+
+
+@patch(
+    "saleor.graphql.order.mutations.fulfillment_update_tracking.send_fulfillment_update"
+)
+def test_fulfillment_update_tracking_send_notification_true(
+    send_fulfillment_update_mock,
+    staff_api_client,
+    fulfillment,
+    permission_manage_orders,
+):
+    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
+    tracking = "stationary tracking"
+    variables = {"id": fulfillment_id, "tracking": tracking, "notifyCustomer": True}
+    response = staff_api_client.post_graphql(
+        FULFILLMENT_UPDATE_TRACKING_QUERY,
+        variables,
+        permissions=[permission_manage_orders],
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["orderFulfillmentUpdateTracking"]["fulfillment"]
+    assert data["trackingNumber"] == tracking
+    send_fulfillment_update_mock.assert_called_once_with(
+        fulfillment.order, fulfillment, ANY
+    )
+
+
+@patch("saleor.order.notifications.send_fulfillment_update")
+def test_fulfillment_update_tracking_send_notification_false(
+    send_fulfillment_update_mock,
+    staff_api_client,
+    fulfillment,
+    permission_manage_orders,
+):
+    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
+    tracking = "stationary tracking"
+    variables = {"id": fulfillment_id, "tracking": tracking, "notifyCustomer": False}
+    response = staff_api_client.post_graphql(
+        FULFILLMENT_UPDATE_TRACKING_QUERY,
+        variables,
+        permissions=[permission_manage_orders],
+    )
+    content = get_graphql_content(response)
+    data = content["data"]["orderFulfillmentUpdateTracking"]["fulfillment"]
+    assert data["trackingNumber"] == tracking
+    send_fulfillment_update_mock.assert_not_called()
