+++ saleor/graphql/order/tests/mutations/test_order_fulfill.py
@@ -6,16 +6,13 @@ import pytest
 from .....core.exceptions import InsufficientStock, InsufficientStockData
 from .....giftcard import GiftCardEvents
 from .....giftcard.models import GiftCard, GiftCardEvent
-from .....order import FulfillmentStatus, OrderEvents, OrderStatus
-from .....order.actions import fulfill_order_lines
+from .....order import FulfillmentStatus, OrderStatus
 from .....order.error_codes import OrderErrorCode
-from .....order.fetch import OrderLineInfo
-from .....order.models import Fulfillment, FulfillmentLine, OrderLine
-from .....plugins.manager import get_plugins_manager
-from .....product.models import Product, ProductVariant
+from .....order.models import Fulfillment, FulfillmentLine
+from .....product.models import ProductVariant
 from .....tests.utils import flush_post_commit_hooks
 from .....warehouse.models import Allocation, Stock
-from ....tests.utils import assert_no_permission, get_graphql_content
+from ....tests.utils import get_graphql_content
 
 ORDER_FULFILL_QUERY = """
     mutation fulfillOrder(
@@ -1410,321 +1407,6 @@ def test_order_fulfill_preorder_waiting_fulfillment(
     )
 
 
-@patch("saleor.plugins.manager.PluginsManager.notify")
-def test_fulfillment_update_tracking(
-    send_fulfillment_update_mock,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-):
-    query = """
-        mutation updateFulfillment($id: ID!, $tracking: String) {
-            orderFulfillmentUpdateTracking(
-                id: $id, input: { trackingNumber: $tracking }
-            ) {
-                fulfillment {
-                    trackingNumber
-                }
-            }
-        }
-    """
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    tracking = "stationary tracking"
-    variables = {"id": fulfillment_id, "tracking": tracking}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentUpdateTracking"]["fulfillment"]
-    assert data["trackingNumber"] == tracking
-    send_fulfillment_update_mock.assert_not_called()
-
-
-FULFILLMENT_UPDATE_TRACKING_WITH_SEND_NOTIFICATION_QUERY = """
-    mutation updateFulfillment(
-        $id: ID!
-        $tracking: String
-        $notifyCustomer: Boolean
-    ) {
-        orderFulfillmentUpdateTracking(
-            id: $id
-            input: { trackingNumber: $tracking, notifyCustomer: $notifyCustomer }
-        ) {
-            fulfillment {
-                trackingNumber
-            }
-        }
-    }
-"""
-
-
-@patch(
-    "saleor.graphql.order.mutations.fulfillment_update_tracking.send_fulfillment_update"
-)
-def test_fulfillment_update_tracking_send_notification_true(
-    send_fulfillment_update_mock,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-):
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    tracking = "stationary tracking"
-    variables = {"id": fulfillment_id, "tracking": tracking, "notifyCustomer": True}
-    response = staff_api_client.post_graphql(
-        FULFILLMENT_UPDATE_TRACKING_WITH_SEND_NOTIFICATION_QUERY,
-        variables,
-        permissions=[permission_manage_orders],
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentUpdateTracking"]["fulfillment"]
-    assert data["trackingNumber"] == tracking
-    send_fulfillment_update_mock.assert_called_once_with(
-        fulfillment.order, fulfillment, ANY
-    )
-
-
-@patch("saleor.order.notifications.send_fulfillment_update")
-def test_fulfillment_update_tracking_send_notification_false(
-    send_fulfillment_update_mock,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-):
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    tracking = "stationary tracking"
-    variables = {"id": fulfillment_id, "tracking": tracking, "notifyCustomer": False}
-    response = staff_api_client.post_graphql(
-        FULFILLMENT_UPDATE_TRACKING_WITH_SEND_NOTIFICATION_QUERY,
-        variables,
-        permissions=[permission_manage_orders],
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentUpdateTracking"]["fulfillment"]
-    assert data["trackingNumber"] == tracking
-    send_fulfillment_update_mock.assert_not_called()
-
-
-CANCEL_FULFILLMENT_MUTATION = """
-    mutation cancelFulfillment($id: ID!, $warehouseId: ID) {
-        orderFulfillmentCancel(id: $id, input: {warehouseId: $warehouseId}) {
-            fulfillment {
-                status
-            }
-            order {
-                status
-            }
-            errors {
-                code
-                field
-            }
-        }
-    }
-"""
-
-
-def test_cancel_fulfillment(
-    staff_api_client, fulfillment, staff_user, permission_manage_orders, warehouse
-):
-    query = CANCEL_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.id)
-    variables = {"id": fulfillment_id, "warehouseId": warehouse_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentCancel"]
-    assert data["fulfillment"]["status"] == FulfillmentStatus.CANCELED.upper()
-    assert data["order"]["status"] == OrderStatus.UNFULFILLED.upper()
-    event_cancelled, event_restocked_items = fulfillment.order.events.all()
-    assert event_cancelled.type == (OrderEvents.FULFILLMENT_CANCELED)
-    assert event_cancelled.parameters == {"composed_id": fulfillment.composed_id}
-    assert event_cancelled.user == staff_user
-
-    assert event_restocked_items.type == (OrderEvents.FULFILLMENT_RESTOCKED_ITEMS)
-    assert event_restocked_items.parameters == {
-        "quantity": fulfillment.get_total_quantity(),
-        "warehouse": str(warehouse.pk),
-    }
-    assert event_restocked_items.user == staff_user
-    assert Fulfillment.objects.filter(
-        pk=fulfillment.pk, status=FulfillmentStatus.CANCELED
-    ).exists()
-
-
-def test_cancel_fulfillment_for_order_with_gift_card_lines(
-    staff_api_client,
-    fulfillment,
-    gift_card_shippable_order_line,
-    staff_user,
-    permission_manage_orders,
-    warehouse,
-):
-    query = CANCEL_FULFILLMENT_MUTATION
-    order = gift_card_shippable_order_line.order
-    order_fulfillment = order.fulfillments.first()
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", order_fulfillment.id)
-    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.id)
-    variables = {"id": fulfillment_id, "warehouseId": warehouse_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentCancel"]
-    assert not data["fulfillment"]
-    assert len(data["errors"]) == 1
-    assert data["errors"][0]["code"] == OrderErrorCode.CANNOT_CANCEL_FULFILLMENT.name
-    assert data["errors"][0]["field"] == "fulfillment"
-
-
-def test_cancel_fulfillment_no_warehouse_id(
-    staff_api_client, fulfillment, permission_manage_orders
-):
-    query = """
-        mutation cancelFulfillment($id: ID!) {
-            orderFulfillmentCancel(id: $id) {
-                fulfillment {
-                    status
-                }
-                order {
-                    status
-                }
-                errors {
-                    code
-                    field
-                }
-            }
-        }
-    """
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    variables = {"id": fulfillment_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    errors = content["data"]["orderFulfillmentCancel"]["errors"]
-    assert len(errors) == 1
-    error = errors[0]
-    assert error["field"] == "warehouseId"
-    assert error["code"] == OrderErrorCode.REQUIRED.name
-
-
-@patch("saleor.order.actions.restock_fulfillment_lines")
-def test_cancel_fulfillment_awaiting_approval(
-    mock_restock_lines, staff_api_client, fulfillment, permission_manage_orders
-):
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-    query = CANCEL_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    variables = {"id": fulfillment_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentCancel"]
-    assert data["fulfillment"] is None
-    mock_restock_lines.assert_not_called()
-    event_cancelled = fulfillment.order.events.get()
-    assert event_cancelled.type == (OrderEvents.FULFILLMENT_CANCELED)
-    assert event_cancelled.parameters == {}
-    assert event_cancelled.user == staff_api_client.user
-    assert not Fulfillment.objects.filter(pk=fulfillment.pk).exists()
-
-
-@patch("saleor.order.actions.restock_fulfillment_lines")
-def test_cancel_fulfillment_awaiting_approval_warehouse_specified(
-    mock_restock_lines,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-    warehouse,
-):
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-    query = CANCEL_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.id)
-    variables = {"id": fulfillment_id, "warehouseId": warehouse_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentCancel"]
-    assert data["fulfillment"] is None
-    mock_restock_lines.assert_not_called()
-    event_cancelled = fulfillment.order.events.get()
-    assert event_cancelled.type == (OrderEvents.FULFILLMENT_CANCELED)
-    assert event_cancelled.parameters == {}
-    assert event_cancelled.user == staff_api_client.user
-    assert not Fulfillment.objects.filter(pk=fulfillment.pk).exists()
-
-
-def test_cancel_fulfillment_canceled_state(
-    staff_api_client, fulfillment, permission_manage_orders, warehouse
-):
-    query = CANCEL_FULFILLMENT_MUTATION
-    fulfillment.status = FulfillmentStatus.CANCELED
-    fulfillment.save(update_fields=["status"])
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.id)
-    variables = {"id": fulfillment_id, "warehouseId": warehouse_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    errors = content["data"]["orderFulfillmentCancel"]["errors"]
-    assert len(errors) == 1
-    error = errors[0]
-    assert error["field"] == "fulfillment"
-    assert error["code"] == OrderErrorCode.CANNOT_CANCEL_FULFILLMENT.name
-
-
-def test_cancel_fulfillment_warehouse_without_stock(
-    order_line, warehouse, staff_api_client, permission_manage_orders, staff_user
-):
-    query = CANCEL_FULFILLMENT_MUTATION
-    order = order_line.order
-    fulfillment = order.fulfillments.create(tracking_number="123")
-    fulfillment.lines.create(order_line=order_line, quantity=order_line.quantity)
-    order.status = OrderStatus.FULFILLED
-    order.save(update_fields=["status"])
-
-    assert not Stock.objects.filter(
-        warehouse=warehouse, product_variant=order_line.variant
-    )
-    assert not Allocation.objects.filter(order_line=order_line)
-
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    warehouse_id = graphene.Node.to_global_id("Warehouse", warehouse.id)
-    variables = {"id": fulfillment_id, "warehouseId": warehouse_id}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentCancel"]["fulfillment"]
-    assert data["status"] == FulfillmentStatus.CANCELED.upper()
-    event_cancelled, event_restocked_items = fulfillment.order.events.all()
-    assert event_cancelled.type == (OrderEvents.FULFILLMENT_CANCELED)
-    assert event_cancelled.parameters == {"composed_id": fulfillment.composed_id}
-    assert event_cancelled.user == staff_user
-
-    assert event_restocked_items.type == (OrderEvents.FULFILLMENT_RESTOCKED_ITEMS)
-    assert event_restocked_items.parameters == {
-        "quantity": fulfillment.get_total_quantity(),
-        "warehouse": str(warehouse.pk),
-    }
-    assert event_restocked_items.user == staff_user
-
-    stock = Stock.objects.filter(
-        warehouse=warehouse, product_variant=order_line.variant
-    ).first()
-    assert stock.quantity == order_line.quantity
-    allocation = order_line.allocations.filter(stock=stock).first()
-    assert allocation.quantity_allocated == order_line.quantity
-
-
 @patch("saleor.order.actions.send_fulfillment_confirmation_to_customer", autospec=True)
 def test_create_digital_fulfillment(
     mock_email_fulfillment,
@@ -1774,542 +1456,3 @@ def test_create_digital_fulfillment(
     get_graphql_content(response)
 
     assert mock_email_fulfillment.call_count == 1
-
-
-APPROVE_FULFILLMENT_MUTATION = """
-    mutation approveFulfillment(
-        $id: ID!, $notifyCustomer: Boolean!, $allowStockToBeExceeded: Boolean = false
-    ) {
-        orderFulfillmentApprove(
-                id: $id,
-                notifyCustomer: $notifyCustomer,
-                allowStockToBeExceeded: $allowStockToBeExceeded) {
-            fulfillment {
-                status
-            }
-            order {
-                status
-            }
-            errors {
-                field
-                code
-                message
-            }
-        }
-    }
-"""
-
-
-@patch("saleor.plugins.manager.PluginsManager.fulfillment_approved")
-@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer", autospec=True)
-def test_fulfillment_approve(
-    mock_email_fulfillment,
-    mock_fulfillment_approved,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-):
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-    query = APPROVE_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    variables = {"id": fulfillment_id, "notifyCustomer": True}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentApprove"]
-    assert not data["errors"]
-    assert data["fulfillment"]["status"] == FulfillmentStatus.FULFILLED.upper()
-    assert data["order"]["status"] == OrderStatus.FULFILLED.upper()
-    fulfillment.refresh_from_db()
-    assert fulfillment.status == FulfillmentStatus.FULFILLED
-
-    assert mock_email_fulfillment.call_count == 1
-    events = fulfillment.order.events.all()
-    assert len(events) == 1
-    event = events[0]
-    assert event.type == OrderEvents.FULFILLMENT_FULFILLED_ITEMS
-    assert event.user == staff_api_client.user
-    mock_fulfillment_approved.assert_called_once_with(fulfillment)
-
-
-@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer", autospec=True)
-def test_fulfillment_approve_delete_products_before_approval_allow_stock_exceeded_true(
-    mock_email_fulfillment,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-):
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-
-    Product.objects.all().delete()
-
-    query = APPROVE_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    variables = {
-        "id": fulfillment_id,
-        "notifyCustomer": True,
-        "allowStockToBeExceeded": True,
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentApprove"]
-    assert not data["errors"]
-    assert data["fulfillment"]["status"] == FulfillmentStatus.FULFILLED.upper()
-    assert data["order"]["status"] == OrderStatus.FULFILLED.upper()
-    fulfillment.refresh_from_db()
-    assert fulfillment.status == FulfillmentStatus.FULFILLED
-
-    assert mock_email_fulfillment.call_count == 1
-    events = fulfillment.order.events.all()
-    assert len(events) == 1
-    event = events[0]
-    assert event.type == OrderEvents.FULFILLMENT_FULFILLED_ITEMS
-    assert event.user == staff_api_client.user
-
-
-@patch("saleor.plugins.manager.PluginsManager.fulfillment_approved")
-@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer", autospec=True)
-def test_fulfillment_approve_delete_products_before_approval_allow_stock_exceeded_false(
-    mock_email_fulfillment,
-    mock_fulfillment_approved,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-):
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-
-    Product.objects.all().delete()
-
-    query = APPROVE_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    variables = {
-        "id": fulfillment_id,
-        "notifyCustomer": True,
-        "allowStockToBeExceeded": False,
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-
-    content = get_graphql_content(response)
-    errors = content["data"]["orderFulfillmentApprove"]["errors"]
-
-    assert len(errors) == 2
-
-    error_field_and_code = {
-        "field": "stocks",
-        "code": "INSUFFICIENT_STOCK",
-    }
-    expected_errors = [
-        {
-            **error_field_and_code,
-            "message": f"Insufficient product stock: {line.order_line}",
-        }
-        for line in fulfillment.lines.all()
-    ]
-
-    for expected_error in expected_errors:
-        assert expected_error in errors
-
-    fulfillment.refresh_from_db()
-    assert fulfillment.status == FulfillmentStatus.WAITING_FOR_APPROVAL
-
-    assert mock_email_fulfillment.call_count == 1
-    events = fulfillment.order.events.all()
-    assert len(events) == 0
-    mock_fulfillment_approved.assert_not_called()
-
-
-@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer", autospec=True)
-def test_fulfillment_approve_gift_cards_created(
-    mock_email_fulfillment,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-    gift_card_shippable_order_line,
-    gift_card_non_shippable_order_line,
-):
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-
-    gift_card_line_1 = gift_card_shippable_order_line
-    gift_card_line_2 = gift_card_non_shippable_order_line
-    stock_1 = gift_card_line_1.variant.stocks.first()
-    stock_2 = gift_card_line_2.variant.stocks.first()
-    fulfillment_line_1 = fulfillment.lines.create(
-        order_line=gift_card_line_1, quantity=gift_card_line_1.quantity, stock=stock_1
-    )
-    fulfillment_line_2 = fulfillment.lines.create(
-        order_line=gift_card_line_2, quantity=gift_card_line_2.quantity, stock=stock_2
-    )
-
-    fulfill_order_lines(
-        [
-            OrderLineInfo(
-                line=gift_card_line_1,
-                quantity=gift_card_line_1.quantity,
-                warehouse_pk=stock_1.warehouse.pk,
-            ),
-            OrderLineInfo(
-                line=gift_card_line_2,
-                quantity=gift_card_line_2.quantity,
-                warehouse_pk=stock_2.warehouse.pk,
-            ),
-        ],
-        manager=get_plugins_manager(),
-    )
-
-    query = APPROVE_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    variables = {"id": fulfillment_id, "notifyCustomer": True}
-    assert GiftCard.objects.count() == 0
-
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentApprove"]
-    assert not data["errors"]
-    assert data["fulfillment"]["status"] == FulfillmentStatus.FULFILLED.upper()
-    assert data["order"]["status"] == OrderStatus.FULFILLED.upper()
-    fulfillment.refresh_from_db()
-    assert fulfillment.status == FulfillmentStatus.FULFILLED
-
-    assert mock_email_fulfillment.call_count == 1
-    events = fulfillment.order.events.all()
-    assert len(events) == 1
-    event = events[0]
-    assert event.type == OrderEvents.FULFILLMENT_FULFILLED_ITEMS
-    assert event.user == staff_api_client.user
-    gift_cards = GiftCard.objects.all()
-    assert gift_cards.count() == gift_card_line_1.quantity + gift_card_line_2.quantity
-    for gift_card in gift_cards:
-        if gift_card.product == gift_card_line_1.variant.product:
-            assert gift_card.fulfillment_line == fulfillment_line_1
-        else:
-            assert gift_card.fulfillment_line == fulfillment_line_2
-
-
-@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer", autospec=True)
-def test_fulfillment_approve_when_stock_is_exceeded_and_flag_enabled(
-    mock_email_fulfillment,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-):
-    # make stocks exceeded
-    for stock in [line.stock for line in fulfillment.lines.all()]:
-        stock.quantity = -99
-        stock.save()
-
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-    query = APPROVE_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-
-    # make response with flag disabled, raised error is expected
-    variables = {
-        "id": fulfillment_id,
-        "notifyCustomer": True,
-        "allowStockToBeExceeded": True,
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentApprove"]
-    assert not data["errors"]
-    assert data["fulfillment"]["status"] == FulfillmentStatus.FULFILLED.upper()
-    assert data["order"]["status"] == OrderStatus.FULFILLED.upper()
-    fulfillment.refresh_from_db()
-    assert fulfillment.status == FulfillmentStatus.FULFILLED
-
-    assert mock_email_fulfillment.call_count == 1
-    events = fulfillment.order.events.all()
-    assert len(events) == 1
-    event = events[0]
-    assert event.type == OrderEvents.FULFILLMENT_FULFILLED_ITEMS
-    assert event.user == staff_api_client.user
-
-
-@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer", autospec=True)
-def test_fulfillment_approve_when_stock_is_exceeded_and_flag_disabled(
-    mock_email_fulfillment,
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-):
-    # make stocks exceeded
-    for stock in [line.stock for line in fulfillment.lines.all()]:
-        stock.quantity = -99
-        stock.save()
-
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-    query = APPROVE_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-
-    variables = {
-        "id": fulfillment_id,
-        "notifyCustomer": True,
-        "allowStockToBeExceeded": False,
-    }
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response, ignore_errors=True)
-    errors = content["data"]["orderFulfillmentApprove"]["errors"]
-
-    assert len(errors) == 2
-
-    error_field_and_code = {
-        "field": "stocks",
-        "code": "INSUFFICIENT_STOCK",
-    }
-
-    expected_errors = [
-        {
-            **error_field_and_code,
-            "message": f"Insufficient product stock: {line.order_line}",
-        }
-        for line in fulfillment.lines.all()
-    ]
-
-    for expected_error in expected_errors:
-        assert expected_error in errors
-
-
-@patch("saleor.plugins.manager.PluginsManager.fulfillment_approved")
-@patch("saleor.order.actions.send_fulfillment_confirmation_to_customer", autospec=True)
-def test_fulfillment_approve_partial_order_fulfill(
-    mock_email_fulfillment,
-    mock_fulfillment_approved,
-    staff_api_client,
-    fulfillment_awaiting_approval,
-    permission_manage_orders,
-):
-    # given
-    query = APPROVE_FULFILLMENT_MUTATION
-    order = fulfillment_awaiting_approval.order
-
-    second_fulfillment = order.fulfillments.create()
-    line_1 = order.lines.first()
-    line_2 = order.lines.last()
-    second_fulfillment.lines.create(
-        order_line=line_1, quantity=line_1.quantity - line_1.quantity_fulfilled
-    )
-    second_fulfillment.lines.create(
-        order_line=line_2, quantity=line_2.quantity - line_2.quantity_fulfilled
-    )
-    second_fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    second_fulfillment.save()
-
-    line_1.quantity_fulfilled = line_1.quantity
-    line_2.quantity_fulfilled = line_2.quantity
-
-    OrderLine.objects.bulk_update([line_1, line_2], ["quantity_fulfilled"])
-
-    fulfillment_id = graphene.Node.to_global_id(
-        "Fulfillment", fulfillment_awaiting_approval.id
-    )
-    variables = {"id": fulfillment_id, "notifyCustomer": False}
-
-    # when
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-
-    # then
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentApprove"]
-    assert not data["errors"]
-    assert data["fulfillment"]["status"] == FulfillmentStatus.FULFILLED.upper()
-    assert data["order"]["status"] == "PARTIALLY_FULFILLED"
-    fulfillment_awaiting_approval.refresh_from_db()
-    assert fulfillment_awaiting_approval.status == FulfillmentStatus.FULFILLED
-
-    assert mock_email_fulfillment.call_count == 0
-    mock_fulfillment_approved.assert_called_once_with(fulfillment_awaiting_approval)
-
-
-def test_fulfillment_approve_invalid_status(
-    staff_api_client,
-    fulfillment,
-    permission_manage_orders,
-):
-    query = APPROVE_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    variables = {"id": fulfillment_id, "notifyCustomer": True}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentApprove"]
-    assert data["errors"][0]["code"] == OrderErrorCode.INVALID.name
-
-
-def test_fulfillment_approve_order_unpaid(
-    staff_api_client,
-    fulfillment,
-    site_settings,
-    permission_manage_orders,
-):
-    site_settings.fulfillment_allow_unpaid = False
-    site_settings.save(update_fields=["fulfillment_allow_unpaid"])
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-    query = APPROVE_FULFILLMENT_MUTATION
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    variables = {"id": fulfillment_id, "notifyCustomer": True}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentApprove"]
-    assert data["errors"][0]["code"] == OrderErrorCode.CANNOT_FULFILL_UNPAID_ORDER.name
-
-
-def test_fulfillment_approve_preorder(
-    staff_api_client, fulfillment, permission_manage_orders, site_settings
-):
-    """Fulfillment with WAITING_FOR_APPROVAL status can not be fulfilled
-    if it contains variant in preorder."""
-    site_settings.fulfillment_auto_approve = False
-    site_settings.save(update_fields=["fulfillment_auto_approve"])
-
-    order_line = fulfillment.order.lines.first()
-    variant = order_line.variant
-    variant.is_preorder = True
-    variant.save(update_fields=["is_preorder"])
-    fulfillment.status = FulfillmentStatus.WAITING_FOR_APPROVAL
-    fulfillment.save(update_fields=["status"])
-    query = APPROVE_FULFILLMENT_MUTATION
-
-    fulfillment_id = graphene.Node.to_global_id("Fulfillment", fulfillment.id)
-    variables = {"id": fulfillment_id, "notifyCustomer": False}
-    response = staff_api_client.post_graphql(
-        query, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["orderFulfillmentApprove"]
-    assert data["errors"]
-
-    error = data["errors"][0]
-    assert error["field"] == "orderLineId"
-    assert error["code"] == OrderErrorCode.FULFILL_ORDER_LINE.name
-
-
-QUERY_FULFILLMENT = """
-    query fulfillment($id: ID!) {
-        order(id: $id) {
-            fulfillments {
-                id
-                fulfillmentOrder
-                status
-                trackingNumber
-                warehouse {
-                    id
-                }
-                lines {
-                    orderLine {
-                        id
-                    }
-                    quantity
-                }
-            }
-        }
-    }
-"""
-
-
-def test_fulfillment_query(
-    staff_api_client,
-    fulfilled_order,
-    warehouse,
-    permission_manage_orders,
-):
-    # given
-    order = fulfilled_order
-    order_line_1, order_line_2 = order.lines.all()
-    order_id = graphene.Node.to_global_id("Order", order.pk)
-    order_line_1_id = graphene.Node.to_global_id("OrderLine", order_line_1.pk)
-    order_line_2_id = graphene.Node.to_global_id("OrderLine", order_line_2.pk)
-    warehose_id = graphene.Node.to_global_id("Warehouse", warehouse.pk)
-    staff_user = staff_api_client.user
-    staff_user.user_permissions.add(permission_manage_orders)
-    variables = {"id": order_id}
-
-    # when
-    response = staff_api_client.post_graphql(QUERY_FULFILLMENT, variables)
-
-    # then
-    content = get_graphql_content(response)
-    data = content["data"]["order"]["fulfillments"]
-    assert len(data) == 1
-    fulfillment_data = data[0]
-
-    assert fulfillment_data["fulfillmentOrder"] == 1
-    assert fulfillment_data["status"] == FulfillmentStatus.FULFILLED.upper()
-    assert fulfillment_data["trackingNumber"] == "123"
-    assert fulfillment_data["warehouse"]["id"] == warehose_id
-    assert len(fulfillment_data["lines"]) == 2
-    assert {
-        "orderLine": {"id": order_line_1_id},
-        "quantity": order_line_1.quantity,
-    } in fulfillment_data["lines"]
-    assert {
-        "orderLine": {"id": order_line_2_id},
-        "quantity": order_line_2.quantity,
-    } in fulfillment_data["lines"]
-
-
-QUERY_ORDER_FULFILL_DATA = """
-    query OrderFulfillData($id: ID!) {
-        order(id: $id) {
-            id
-            lines {
-                variant {
-                    stocks {
-                        warehouse {
-                            id
-                        }
-                        quantity
-                        quantityAllocated
-                    }
-                }
-            }
-        }
-    }
-"""
-
-
-def test_staff_can_query_order_fulfill_data(
-    staff_api_client, order_with_lines, permission_manage_orders
-):
-    order_id = graphene.Node.to_global_id("Order", order_with_lines.pk)
-    variables = {"id": order_id}
-    response = staff_api_client.post_graphql(
-        QUERY_ORDER_FULFILL_DATA, variables, permissions=[permission_manage_orders]
-    )
-    content = get_graphql_content(response)
-    data = content["data"]["order"]["lines"]
-    assert len(data) == 2
-    assert data[0]["variant"]["stocks"][0]["quantity"] == 5
-    assert data[0]["variant"]["stocks"][0]["quantityAllocated"] == 3
-    assert data[1]["variant"]["stocks"][0]["quantity"] == 2
-    assert data[1]["variant"]["stocks"][0]["quantityAllocated"] == 2
-
-
-def test_staff_can_query_order_fulfill_data_without_permission(
-    staff_api_client, order_with_lines
-):
-    order_id = graphene.Node.to_global_id("Order", order_with_lines.pk)
-    variables = {"id": order_id}
-    response = staff_api_client.post_graphql(QUERY_ORDER_FULFILL_DATA, variables)
-    assert_no_permission(response)
